Add security-hardening to cloud-config for instances
I'm following the security hardening I did for the Ubuntu server in
Assignment 2: Virtual Servers with OpenStack. Without unattended-upgrade
Automatic-Reboot (as I don't want to risk Jenkins breaking due to
restarts) and fail2ban (as I don't have time to figure out automated
configuration of it).

With just package_upgrade, "System restart required" was shown after I
logged in. So set it to reboot if required.

sshd_config.d only exists on Debian-like distros, and the ubuntu user
is only the default user on Ubuntu images. So now we should tell the
user of the template that Ubuntu really is required.

Jenkins and/or Salt might require other SSH users, but I'll cross that
bridge when I get there. I can [use OS::Heat::MultipartMime to combine
multiple cloud-configs][1] to use on a single instance.

I'm not exactly sure of the difference between SOFTWARE_CONFIG and RAW,
but SOFTWARE_CONFIG is used in the [Heat documentation][1].

[1]: https://docs.openstack.org/heat/yoga/template_guide/software_deployment.html#software-config-resources