Fix Jenkins install script not installing Jenkins
Trying to install the JDK and Jenkins with the same command results in
the following error:

Errors were encountered while processing:
 jenkins
needrestart is being skipped since dpkg has failed
E: Sub-process /usr/bin/dpkg returned an error code (1)

I'm assuming this is because Jenkins depends on the JDK but
doesn't/isn't able to list this in APT depends, so the JDK isn't fully
installed when apt-get tries to install Jenkins.

Hopefully this isn't too hard to do in cloud-init, to install one
package after another is finished installing.