Prepare Django settings for production
Following Django's [Deployment checklist][1].

I was originally going to use Docker Compose to load the secret key from
a file, but after reading about [SECRET_KEY][2] I realize that it's only
used for things that I'm not using. So I won't bother.

[ALLOWED_HOSTS can't be empty without DEBUG = True][3], and I don't know
what to set it to so I'm using the wildcard.

[1]: https://docs.djangoproject.com/en/4.2/howto/deployment/checklist/
[2]: https://docs.djangoproject.com/en/4.2/ref/settings/#std-setting-SECRET_KEY
[3]: https://docs.djangoproject.com/en/4.2/howto/deployment/checklist/#environment-specific-settings