Clean workspace in first stage of Jenkinsfile
After adding a clean before clone "Additional Behaviour" in the web
interface, it started building again. And even after removing that, it
kept building.

I assume what happened is that because the file still existed once and
prevented builds from completing, the cleanup step I added later was
never run. So just move it to the beginning so it can always run before
the build crashes.